Move AutocorrectFlusher declaration into LSPTypechecker header file
