+++ main/lsp/LSPTypechecker.h
@@ -180,5 +180,15 @@ public:
     const core::GlobalState &state() const;
 };
 
+class AutocorrectFlusher : public core::ErrorFlusher {
+public:
+    std::vector<std::unique_ptr<core::Error>> collectedErrors;
+    AutocorrectFlusher() = default;
+    ~AutocorrectFlusher() = default;
+
+    void flushErrors(spdlog::logger &logger, std::vector<std::unique_ptr<core::ErrorQueueMessage>> errors,
+                     const core::GlobalState &gs, core::FileRef file) override;
+};
+
 } // namespace sorbet::realmain::lsp
 #endif
